Initial cut at functioning world support.
 * Fully support the world parameter in Server methods to mutate permissions.
 * Bless 'default' as signifying the default world/context in the Server
   interface.
 * Load all contexts into a contextMap, mapping contextName -> context.
 * Move functions to delete and update permissions (including groups) into
   the Server base class, in preparation for the next step.

The code at this stage issues what I imagined might be the correct commands to
configure distinct worlds in LuckPerms. Commands pertaining to the default
context don't have a trailing `world=` parameter, whereas those commands
pertaining to a specific world do have that.

Unfortunately, that's not how LuckPerms actually handles world-specific
permissions. It seems that, in the LuckPerms model, groups exist in all worlds,
but have different properties depending on the world. Specifically a group can
have world-specific:

 * parents,
 * weight,
 * permission nodes.

But you can't create a world-specific group. You must omit the `world=`
parameter when creating the group.

For the LuckPerms YAML storage, at least, if you specify world-specific
properties of a group (weight, parents, permissions), then the YAML file always
contains a world-specific entry for that property, even if it's exactly the
same as the property in the default context. That would be a workable solution,
but it's not ideal from the standpoint of readability. So the goal of subsequent
commits is to pare down the LuckPerms commands issued to just those that set
world-specific properties that differ from the default context. Then, looking
at the LuckPerms YAML storage for a given group, you will only see the default
context and the differences between the default and specific worlds.